requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Werkzeug==2.3.7
WTForms==3.0.1
Flask-WTF==1.1.1
Pillow==10.0.1
requests==2.31.0
email-validator==2.0.0
python-dotenv==1.0.0
